✨ Feat: 결제 기능 구현

+++ src/main/java/io/hhplus/concert/hhplusconcert/domain/model/Seat.java
@@ -14,7 +14,7 @@ public record Seat(
         int seatNo,
         SeatStatus status,
         LocalDateTime reservationAt,
-        int seatPrice
+        Long seatPrice
 ) {
     public void checkStatus() {
         if (status.equals(SeatStatus.UNAVAILABLE)) {
